requirements.txt
requests
openai
pyyaml
python-dotenv
crewai
